move the unit test from Hive packge to sql/core
