+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/CachedTableSuite.scala
@@ -203,18 +203,4 @@ class CachedTableSuite extends QueryTest {
     sql("DROP TABLE refreshTable")
     Utils.deleteRecursively(tempPath)
   }
-
-  test("SPARK-10327 Cache Table is not working while subquery has alias in its project list") {
-    import org.apache.spark.sql.hive.execution.HiveTableScan
-    sql("select key, value, key + 1 from src").registerTempTable("abc")
-    cacheTable("abc")
-
-    val sparkPlan = sql(
-      """select a.key, b.key, c.key from
-        |abc a join abc b on a.key=b.key
-        |join abc c on a.key=c.key""".stripMargin).queryExecution.sparkPlan
-
-    assert(sparkPlan.collect { case e: InMemoryColumnarTableScan => e }.size === 3)
-    assert(sparkPlan.collect { case e: HiveTableScan => e }.size === 0)
-  }
 }
